Traffic-Manager-Failover/Python-Flask/requirements.txt
Flask==2.0.2
azure-mgmt-search==8.0
azure-search-documents==11.3.0
